+++ src/app/json-schema-example/json-schema-example.component.ts
@@ -14,6 +14,7 @@ import binding_sample_schema from './binding_sample_schema.json';
 import binding_sample_model from './binding_sample_model.json';
 import binding_sample_bindings from './binding_sample_bindings';
 import visibility_binding_example from './visibility-binding-example-schema.json';
+import visibility_binding_example2 from './visibility-binding-example-schema2.json';
 
 import {AppService, AppData} from '../app.service';
 
@@ -38,6 +39,7 @@ export class JsonSchemaExampleComponent implements OnInit, OnDestroy {
     {label: 'Sample 2 - Custom bindings', event: this.changeSchemaWithBindings, selected: false},
     {label: 'Sample 3 - Otherschema', event: this.changeSchemaOtherschema, selected: false},
     {label: 'Sample 4 - Visibility binding', event: this.changeSchemaVisibilityBinding, selected: false},
+    {label: 'Sample 5 - Visibility binding 2', event: this.changeSchemaVisibilityBinding2, selected: false}
   ];
 
   constructor(
@@ -205,6 +207,14 @@ export class JsonSchemaExampleComponent implements OnInit, OnDestroy {
 
   changeSchemaVisibilityBinding() {
     this.schema = visibility_binding_example;
+    this.model = {};
+    this.fieldBindings = {};
+    this.fieldValidators = {};
+    this.actions = {};
+  }
+
+  changeSchemaVisibilityBinding2() {
+    this.schema = visibility_binding_example2;
     this.model = {};
     this.fieldBindings = {};
     this.fieldValidators = {};